fix warning, type mismatch
/tmp/raw_pdb/src/Examples/ExampleTypes.cpp: In function ‘uint8_t GetLeafSize(PDB::CodeView::TPI::TypeRecordKind)’:
/tmp/raw_pdb/src/Examples/ExampleTypes.cpp:42:37: warning: format ‘%x’ expects argument of type ‘unsigned int’, but argument 2 has type ‘PDB::CodeView::TPI::TypeRecordKind’ [-Wformat=]
   42 |                 printf("Error! 0x%04x bogus type encountered, aborting...\n", kind);
      |                                  ~~~^                                         ~~~~
      |                                     |                                         |
      |                                     unsigned int                              PDB::CodeView::TPI::TypeRecordKind

+++ src/Examples/ExampleTypes.cpp
@@ -39,7 +39,7 @@ uint8_t GetLeafSize(PDB::CodeView::TPI::TypeRecordKind kind)
 		return sizeof(PDB::CodeView::TPI::TypeRecordKind) + sizeof(uint64_t);
 
 	default:
-		printf("Error! 0x%04x bogus type encountered, aborting...\n", kind);
+		printf("Error! 0x%04x bogus type encountered, aborting...\n", (unsigned) kind);
 	}
 	return 0;
 }
